# faster png decode