--- a/tests/test_png.nim
+++ b/tests/test_png.nim
@@ -1,17 +1,10 @@
 import pixie, pixie/fileformats/png, pngsuite, strformat
 
-# for file in pngSuiteFiles:
-#   let
-#     original = cast[seq[uint8]](
-#       readFile(&"tests/images/png/pngsuite/{file}.png")
-#     )
-#     decoded = decodePng(original)
-#     encoded = encodePng(decoded)
-#     decoded2 = decodePng(cast[seq[uint8]](encoded))
-
-#   doAssert decoded.height == decoded2.height
-#   doAssert decoded.width == decoded2.width
-#   doAssert decoded.data == decoded2.data
+for file in pngSuiteFiles:
+  let
+    original = readFile(&"tests/fileformats/png/pngsuite/{file}.png")
+    decoded = decodePng(original)
+    encoded = encodePng(decoded)
 
 block:
   for channels in 1 .. 4: