--- a/tests/validate_png.nim
+++ b/tests/validate_png.nim
@@ -3,7 +3,7 @@ import chroma, pixie/fileformats/png, pngsuite, stb_image/read as stbi, strforma
 for file in pngSuiteFiles:
   let
     data = readFile(&"tests/fileformats/png/pngsuite/{file}.png")
-    pixieLoaded = decodePngRaw(data)
+    pixieLoaded = decodePng(data)
 
   var
     width, height, channels: int