… tests/benchmark_drawCorrect_vs_drawUber.nim
… tests/benchmark_images_blur.nim
… tests/benchmark_images_loop.nim