png.nim

before:
```
pixie decode ....................... 4.125 ms      4.167 ms    ±0.031  x1000
```
after:
```
pixie decode ....................... 3.690 ms      3.722 ms    ±0.026  x1000
```